fix: Clean up toast stories to have better examples
