--- modules/react/toast/stories/examples/WithActionLinkAndCloseIcon.tsx
@@ -1,20 +1,0 @@
-import React from 'react';
-
-import {Toast} from '@workday/canvas-kit-react/toast';
-import {checkIcon} from '@workday/canvas-system-icons-web';
-import {colors} from '@workday/canvas-kit-react/tokens';
-
-export const WithActionLinkAndCloseIcon = () => {
-  const handleClose = () => console.log('close button clicked');
-
-  return (
-    <Toast mode="dialog" aria-label="notification">
-      <Toast.Icon icon={checkIcon} color="greenApple400" />
-      <Toast.Body>
-        <Toast.Message>Your workbook was successfully</Toast.Message>
-        <Toast.Link href="#hyperlink">Custom Link</Toast.Link>
-      </Toast.Body>
-      <Toast.CloseIcon aria-label="Close" onClick={handleClose} />
-    </Toast>
-  );
-};
