+++ modules/react/toast/stories/toast.stories.mdx
@@ -3,10 +3,7 @@ import {SymbolDoc} from '@workday/canvas-kit-docs';
 import {Toast} from '@workday/canvas-kit-react/toast';
 import {Basic} from './examples/Basic';
 import {ToastAlert} from './examples/ToastAlert';
-import {WithCloseButton} from './examples/WithCloseButton';
 import {ToastDialog} from './examples/ToastDialog';
-import {WithActionLinkAndCloseIcon} from './examples/WithActionLinkAndCloseIcon';
-import {WithPopper} from './examples/WithPopper';
 import {RTL} from './examples/RTL';
 
 <Meta title="Components/Popups/Toast" component={Toast} />
@@ -41,11 +38,6 @@ such as a successful action.
 A `status` `Toast` will wrap the message within a `polite` ARIA live region with a `role` of
 `status`.
 
-Here's a more complete example with a button which triggers a dismissable `Toast`. The `Toast` is
-positioned using `Popper`.
-
-<ExampleCodeBlock code={WithPopper} />
-
 ### Alert
 
 Set the `mode` prop to `alert` to convey urgent and important information such as an error.
@@ -58,22 +50,19 @@ An `alert` `Toast` will wrap the message within an `assertive` ARIA live region 
 ### Dialog
 
 Set the `mode` prop to `dialog` to convey the presence of a follow-up action. If you use this
-`mode`, you need to add an `aria-label`. This `aria-label` should be like additional information for the
+`mode`, you need to add an `aria-label`. This `aria-label` should be additional information for the
 `Toast`. The `Toast` will also add an `aria-describedby` that links the `Toast` to `Toast.Message`
 so that it is read out to screen readers. The `aria-label` should be different that the contents of
 the `Toast` so there is no duplication being read out by screen readers.
 
 <ExampleCodeBlock code={ToastDialog} />
 
-The `Toast` will be given a role of `dialog`. You need to provide an `aria-label` to add
-additional information regarding the contents.
-
-Screen readers will read it out in the following order: `"Notifcation: Your workbook was
-successfully processed."
+Screen readers will read it out in the following order: "Notifcation: Your workbook was successfully
+processed."
 
 > **Note**: You must use the `Toast.Message` subcomponent when the `mode` is `dialog` so that `id`
 > is tied to the message for accessibility. You can also pass in a `model` with `useToastModel` to
-> provide a specific `id` for the `Toast.Message`:
+> provide a specific `id` for the `Toast.Message`
 
 ```tsx
 export const CustomIDToast = () => {
@@ -92,17 +81,6 @@ export const CustomIDToast = () => {
   );
 };
 ```
-
-### Close Icon Button
-
-Include the `Toast.CloseIcon` subcomponent to display a close button within the Toast.
-
-<ExampleCodeBlock code={WithCloseButton} />
-
-`Toast.CloseIcon` may be included within a `dialog` `Toast` to create a `Toast` with both an action
-link and a close button.
-
-<ExampleCodeBlock code={WithActionLinkAndCloseIcon} />
 
 ### Right-to-Left (RTL)
